update model has been repaired
We have separate BlockState object tracking the state of every block in the system; this also defines state update and thus block behavior.  Methods for having the state update itself according to neighbors and retrieving power emissions.

Simulation is responsible for setting off update propagations.  Blocks that update and change expect to propagate update commands to neighbors unless neighbors are schedulable in which case neighbor updates should be scheduled for beginning of next cycle.

BlockState is responsible for update behavior, but update propagation is centralized in Simulation for step scheduling and controllability purposes.